Filter individual export tasks based on their specific associated album
Previously, a task was excluded when its album was part of any of the included/excluded albums.
Now, only the task related to said album is excluded, keeping the other tasks as they are.

Thus, it is now possible to exclude specific albums (e.g. 'WhatsApp') but still import the
images if they are part of any other album as well.